--- VillaBeef/src/main/java/com/villabeef/model/dto/GerenciamentoContas.java
@@ -1,42 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package com.villabeef.model.dto;
-
-
-public class GerenciamentoContas {
-    private String data;
-    private char tipo; //e = entrada; s = saida;
-    private double valor;
-
-    public GerenciamentoContas(String data, char tipo, double valor) {
-        this.data = data;
-        this.tipo = tipo;
-        this.valor = valor;
-    }
-
-    public void setData(String data) {
-        this.data = data;
-    }
-
-    public void setTipo(char tipo) {
-        this.tipo = tipo;
-    }
-
-    public void setValor(double valor) {
-        this.valor = valor;
-    }
-
-    public String getData() {
-        return data;
-    }
-
-    public char getTipo() {
-        return tipo;
-    }
-
-    public double getValor() {
-        return valor;
-    }
-}
